MYBEAUTIP-872 이벤트 수정 API

+++ src/main/java/com/jocoos/mybeautip/client/aws/s3/AwsS3Handler.java
@@ -17,7 +17,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.jocoos.mybeautip.global.constant.MybeautipConstant.TEST_FILE;
-import static com.jocoos.mybeautip.global.util.FileUtil.getFilename;
+import static com.jocoos.mybeautip.global.util.FileUtil.getFileName;
 import static com.jocoos.mybeautip.global.util.ImageUrlConvertUtil.getUri;
 
 @Component
@@ -62,7 +62,7 @@ public class AwsS3Handler {
             return "";
         }
 
-        String filename = getFilename(fileDto.getUrl());
+        String filename = getFileName(fileDto.getUrl());
 
         String path = service.copy(
                 UrlDirectory.TEMP.getDirectory() + filename,
@@ -76,7 +76,7 @@ public class AwsS3Handler {
             return null;
         }
 
-        String filename = getFilename(fileUrl);
+        String filename = getFileName(fileUrl);
 
         // FIXME Test Code 에서 올라오는 File URL 은 실제 파일이 없기때문에 복사할 수가 없어서 회피용 코드 삽입. 뭔가 이 방법 말고 좋은 방법이 있다면... 고쳐주세요
         if (filename.equals(TEST_FILE)) {
@@ -126,6 +126,14 @@ public class AwsS3Handler {
 
     public void delete(FileDto fileDto) {
         String filename = getUri(fileDto.getUrl());
+        service.delete(filename);
+    }
+
+    public void delete(String fileUrl) {
+        if (StringUtils.isBlank(fileUrl)) {
+            return;
+        }
+        String filename = getUri(fileUrl);
         service.delete(filename);
     }
 }
